Add dedicated consumer for order message errors
