+++ src/test/java/uk/gov/companieshouse/itemhandler/kafka/OrderMessageConsumerIntegrationTest.java
@@ -9,7 +9,6 @@ import email.email_send;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
-import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import org.apache.commons.io.IOUtils;
